airflow/requirements.txt
yfinance==0.2.59
pandas==2.2.0
pyyaml==6.0.2